Refactor: Remove unused code and styles from DeptDetailStyles.js and DeptDetailDeptInfoComponentStyles.js
